+++ src/ui/components/Post/substitutes.ts
@@ -1,5 +1,5 @@
 import { Blockquote } from "@components/Blockquote";
-import { InlineCode } from "@components/InlineCode";
+import { InlineCode } from "@components/Code";
 import { Highlight } from "@components/Highlight";
 import { LinkProxy } from "@components/LinkProxy";
 import { CodeBlock } from "@components/CodeBlock";
